Add snapshot testing for queries.
