+++ graphql-parser/Cargo.toml
@@ -20,3 +20,4 @@ thiserror = "1.0.11"
 
 [dev-dependencies]
 pretty_assertions = "0.5.0"
+insta = "0.16.0"
