+++ graphql-parser/tests/queries/fail_bad_args.graphql
@@ -1,0 +1,3 @@
+query MyQuery {
+  field1([something])
+}
