+++ graphql-parser/tests/queries/fail_querry.graphql
@@ -1,0 +1,3 @@
+querry MyQuery {
+  field1
+}
